+++ internal/database/user_input.sql.go
@@ -28,7 +28,7 @@ type CreateUserInputParams struct {
 	Weight        int32
 	DesiredWeight sql.NullInt32
 	TimeFrame     sql.NullInt32
-	Bmi           int32
+	Bmi           float64
 	Program       string
 	CurrKcal      int32
 	Deficit       sql.NullInt32
